update according to fei's comments 1) remove quotavalidationplugin implementation 2) add interface class
