+++ virtualcluster/pkg/syncer/resources/pod/validationplugin/interface.go
@@ -1,0 +1,39 @@
+/*
+Copyright 2022 The Kubernetes Authors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package validationplugin
+
+import (
+	mc "sigs.k8s.io/cluster-api-provider-nested/virtualcluster/pkg/util/mccontroller"
+	uplugin "sigs.k8s.io/cluster-api-provider-nested/virtualcluster/pkg/util/plugin"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sync"
+)
+
+type ValidationPluginInterface interface {
+	Validation(client.Object, string) bool
+	GetTenantLocker(string) *Tenant
+	Enabled() bool
+	ContextInit(*mc.MultiClusterController, bool)
+}
+
+type Tenant struct {
+	ClusterName string
+	Cond        *sync.Mutex
+}
+
+const (
+	QuotaValidationPluginName = "quota"
+)
+
+var ValidationRegister uplugin.ResourceRegister
